Update the tool gamer_extract_uniform

+++ tool/Analysis/gamer_extract_uniform/GAMER_Functions/Interpolation/Int_CQuartic.cpp
@@ -70,7 +70,7 @@ void Int_CQuartic( real CData[], const int CSize[3], const int CStart[3], const 
    real *TDataX = new real [ (CRange[2]+2*CGhost)*TdzX ];   // temporary array after x interpolation
    real *TDataY = new real [ (CRange[2]+2*CGhost)*TdzY ];   // temporary array after y interpolation
 
-   int Idx_InL2, Idx_InL1, Idx_InC, Idx_InR1, Idx_InR2, Idx_Out;
+   long Idx_InL2, Idx_InL1, Idx_InC, Idx_InR1, Idx_InR2, Idx_Out;
    real LSlopeDh_4, RSlopeDh_4, SlopeDh_4, Sign;
 
 
@@ -84,7 +84,7 @@ void Int_CQuartic( real CData[], const int CSize[3], const int CStart[3], const 
          for (int j=CStart[1]-CGhost;    j<CStart[1]+CRange[1]+CGhost;  j++)
          for (int i=CStart[0]-CGhost+1;  i<CStart[0]+CRange[0]+CGhost;  i++)
          {
-            Idx_InC       = k*Cdz + j*Cdy + i*Cdx;
+            Idx_InC       = (long)k*Cdz + j*Cdy + i*Cdx;
             Idx_InL1      = Idx_InC - Cdx;
             CPtr[Idx_InC] = ELBDM_UnwrapPhase( CPtr[Idx_InL1], CPtr[Idx_InC] );
          }
@@ -97,8 +97,8 @@ void Int_CQuartic( real CData[], const int CSize[3], const int CStart[3], const 
       for (int In_y=CStart[1]-CGhost, Out_y=0;  In_y<CStart[1]+CRange[1]+CGhost;  In_y++, Out_y++)
       for (int In_x=CStart[0],        Out_x=0;  In_x<CStart[0]+CRange[0];         In_x++, Out_x+=2)
       {
-         Idx_Out  = Out_z*TdzX + Out_y*Tdy + Out_x*Tdx;
-         Idx_InC  =  In_z*Cdz  +  In_y*Cdy +  In_x*Cdx;
+         Idx_Out  = (long)Out_z*TdzX + Out_y*Tdy + Out_x*Tdx;
+         Idx_InC  = (long) In_z*Cdz  +  In_y*Cdy +  In_x*Cdx;
          Idx_InL1 = Idx_InC  - Cdx;
          Idx_InL2 = Idx_InL1 - Cdx;
          Idx_InR1 = Idx_InC  + Cdx;
@@ -144,7 +144,7 @@ void Int_CQuartic( real CData[], const int CSize[3], const int CStart[3], const 
          for (int j=1;  j<CRange[1]+2*CGhost;  j++)
          for (int i=0;  i<2*CRange[0];         i++)
          {
-            Idx_InC         = k*TdzX + j*Tdy + i*Tdx;
+            Idx_InC         = (long)k*TdzX + j*Tdy + i*Tdx;
             Idx_InL1        = Idx_InC - Tdy;
             TDataX[Idx_InC] = ELBDM_UnwrapPhase( TDataX[Idx_InL1], TDataX[Idx_InC] );
          }
@@ -157,8 +157,8 @@ void Int_CQuartic( real CData[], const int CSize[3], const int CStart[3], const 
       for (int In_y=CGhost, Out_y=0;  In_y   <CGhost+CRange[1];    In_y++, Out_y+=2)
       for (int InOut_x=0;             InOut_x<2*CRange[0];         InOut_x++)
       {
-         Idx_Out  = InOut_z*TdzY + Out_y*Tdy + InOut_x*Tdx;
-         Idx_InC  = InOut_z*TdzX +  In_y*Tdy + InOut_x*Tdx;
+         Idx_Out  = (long)InOut_z*TdzY + Out_y*Tdy + InOut_x*Tdx;
+         Idx_InC  = (long)InOut_z*TdzX +  In_y*Tdy + InOut_x*Tdx;
          Idx_InL1 = Idx_InC  - Tdy;
          Idx_InL2 = Idx_InL1 - Tdy;
          Idx_InR1 = Idx_InC  + Tdy;
@@ -203,7 +203,7 @@ void Int_CQuartic( real CData[], const int CSize[3], const int CStart[3], const 
          for (int j=0;  j<2*CRange[1];         j++)
          for (int i=0;  i<2*CRange[0];         i++)
          {
-            Idx_InC         = k*TdzY + j*Tdy + i*Tdx;
+            Idx_InC         = (long)k*TdzY + j*Tdy + i*Tdx;
             Idx_InL1        = Idx_InC - TdzY;
             TDataY[Idx_InC] = ELBDM_UnwrapPhase( TDataY[Idx_InL1], TDataY[Idx_InC] );
          }
@@ -216,8 +216,8 @@ void Int_CQuartic( real CData[], const int CSize[3], const int CStart[3], const 
       for (int In_y=0,      Out_y=FStart[1];  In_y<2*CRange[1];       In_y++, Out_y++)
       for (int In_x=0,      Out_x=FStart[0];  In_x<2*CRange[0];       In_x++, Out_x++)
       {
-         Idx_Out  = Out_z*Fdz  + Out_y*Fdy + Out_x*Fdx;
-         Idx_InC  =  In_z*TdzY +  In_y*Tdy +  In_x*Tdx;
+         Idx_Out  = (long)Out_z*Fdz  + Out_y*Fdy + Out_x*Fdx;
+         Idx_InC  = (long) In_z*TdzY +  In_y*Tdy +  In_x*Tdx;
          Idx_InL1 = Idx_InC  - TdzY;
          Idx_InL2 = Idx_InL1 - TdzY;
          Idx_InR1 = Idx_InC  + TdzY;
